+++ src/Resources/contao/classes/ModuleVisitorStatPageCounter.php
@@ -50,7 +50,7 @@ class ModuleVisitorStatPageCounter extends \BackendModule
         parent::__construct();
 
         $this->today     = date('Y-m-d');
-        $this->yesterday = date('Y-m-d', mktime(0, 0, 0, date("m"), date("d")-1, date("Y")));
+        $this->yesterday = date('Y-m-d', mktime(0, 0, 0, (int) date("m"), (int) date("d")-1, (int) date("Y")));
     }
 
     protected function compile()
@@ -291,7 +291,7 @@ class ModuleVisitorStatPageCounter extends \BackendModule
     public function generatePageVisitHitDays($VisitorsID, $limit=20, $days=7)
     {
         $arrPageStatCount = false;
-        $week = date('Y-m-d', mktime(0, 0, 0, date("m"), date("d")-$days, date("Y")));
+        $week = date('Y-m-d', mktime(0, 0, 0, (int) date("m"), (int) date("d")-$days, (int) date("Y")));
 
         $this->TemplatePartial = new \BackendTemplate('mod_visitors_be_stat_partial_pagevisithitdays');
 
@@ -495,7 +495,7 @@ class ModuleVisitorStatPageCounter extends \BackendModule
      */
     public function generatePageVisitHitTopDays($VisitorsID, $days = 365, $parse = false)
     {
-        $STARTDATE = date("Y-m-d", mktime(0, 0, 0, date("m"), date("d")-$days, date("Y"))); 
+        $STARTDATE = date("Y-m-d", mktime(0, 0, 0, (int) date("m"), (int) date("d")-$days, (int) date("Y"))); 
         $arrPageStatCount = false;
         $objPageStatCount = \Database::getInstance()
                             ->prepare("SELECT
